chore(vmm): Warn on unexpected subleaf for CPUID leaf 0xB
KVM no longer returns any subleaves greater than 0. The patch was merged
in v6.2 and backported to v5.10. So for all our supported kernels,
subleaves >= 2 should not be included [1].

However, we intentionally leave Firecracker not fail for unsupported
kernels to keep working. Note that we can detect KVM regression thanks
to the test that compares a fingerprint with its baseline.

So prints a warning message without bailing out.

[1]: https://github.com/torvalds/linux/commit/45e966fcca03ecdcccac7cb236e16eea38cc18af

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>